projet/requirements.txt
Flask==2.0.1
mysql_connector_repackaged==0.3.1
pip==20.0.2
psutil==5.9.0
py_cpuinfo==8.0.0
requests==2.22.0
